refactor: Queries to Mutations
