+++ frontend/src/graphql/queries.js
@@ -22,43 +22,11 @@ export const logout = gql`
   }
 `
 
-export const resetPassword = gql`
-  query($sessionId: Float!, $email: String!, $password: String!) {
-    resetPassword(sessionId: $sessionId, email: $email, password: $password)
-  }
-`
-
 export const loginViaEmailVerificationCode = gql`
   query($optin: String!) {
     loginViaEmailVerificationCode(optin: $optin) {
       sessionId
       email
-    }
-  }
-`
-
-export const updateUserInfos = gql`
-  query(
-    $email: String!
-    $firstName: String
-    $lastName: String
-    $description: String
-    $username: String
-    $password: String
-    $passwordNew: String
-    $locale: String
-  ) {
-    updateUserInfos(
-      email: $email
-      firstName: $firstName
-      lastName: $lastName
-      description: $description
-      username: $username
-      password: $password
-      passwordNew: $passwordNew
-      language: $locale
-    ) {
-      validValues
     }
   }
 `
