+++ frontend/src/views/Pages/UserProfile/UserCard_FormUserMail.spec.js
@@ -21,7 +21,7 @@ describe('UserCard_FormUserMail', () => {
       },
     },
     $apollo: {
-      query: mockAPIcall,
+      mutate: mockAPIcall,
     },
   }
