+++ frontend/src/views/Pages/UserProfile/UserCard_FormUsername.spec.js
@@ -35,7 +35,7 @@ describe('UserCard_FormUsername', () => {
       error: toastErrorMock,
     },
     $apollo: {
-      query: mockAPIcall,
+      mutate: mockAPIcall,
     },
   }
