+++ frontend/src/views/Pages/UserProfile/UserCard_FormUsername.vue
@@ -67,7 +67,7 @@
   </b-card>
 </template>
 <script>
-import { updateUserInfos } from '../../../graphql/queries'
+import { updateUserInfos } from '../../../graphql/mutations'
 
 export default {
   name: 'FormUsername',
@@ -87,8 +87,8 @@ export default {
     },
     async onSubmit() {
       this.$apollo
-        .query({
-          query: updateUserInfos,
+        .mutate({
+          mutation: updateUserInfos,
           variables: {
             email: this.$store.state.email,
             username: this.form.username,
