+++ frontend/src/views/Pages/UserProfile/UserCard_Language.spec.js
@@ -25,7 +25,7 @@ describe('UserCard_Language', () => {
       error: toastErrorMock,
     },
     $apollo: {
-      query: mockAPIcall,
+      mutate: mockAPIcall,
     },
   }
