+++ frontend/src/views/Pages/UserProfile/UserCard_Language.vue
@@ -63,7 +63,7 @@
 </template>
 <script>
 import LanguageSwitchSelect from '../../../components/LanguageSwitchSelect.vue'
-import { updateUserInfos } from '../../../graphql/queries'
+import { updateUserInfos } from '../../../graphql/mutations'
 
 export default {
   name: 'FormUserLanguage',
@@ -90,8 +90,8 @@ export default {
 
     async onSubmit() {
       this.$apollo
-        .query({
-          query: updateUserInfos,
+        .mutate({
+          mutation: updateUserInfos,
           variables: {
             language: this.$store.state.language,
           },
